requirements.txt
asgiref==3.10.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.11.12
cffi==2.0.0
channels==4.3.1
channels-redis==4.2.0
charset-normalizer==3.4.3
constantly==23.10.4
cryptography==46.0.1
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-cors-headers==4.9.0
django-environ==0.11.2
django-filter==23.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
gunicorn==23.0.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
msgpack==1.1.2
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyOpenSSL==25.3.0
python-decouple==3.8
PyYAML==6.0.3
qrcode==8.2
redis==5.0.1
referencing==0.37.0
reportlab==4.4.4
requests==2.31.0
rpds-py==0.29.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
zope.interface==8.0.1
